Create senior citizenship in java
Given gender and age of a person, tell if the person is SENIOR CITIZEN or not. A male with age 60 or more years is Senior Citizen. A female with age 58 or more is Senior Citizen.

Input Format

One line of input containing a alphabet G in uppercase or lowercase (Gender of the person) and a number N (Age of the person) separated by space.