Create comparison 5 submission in java
The comparison is as: A = B, A < B, A > B

Check whether it is Right or Wrong

Input Format

The first line contains the number of test cases t.

Then there are t lines each containing a test case. Each test case contains an Integer followed by a comparison Operator which is again followed by an Integer.