Create mathematical expression in java
The expression will be one of the following expressions:

A + B = C, A - B = C and A * B = C.

where A, B, C are three numbers, S is the sign between A and B, and Q is the '=' sign.

If the expression is right print "Yes", Otherwise print the right answer of the expression.

Input Format

The first line contains the number of test cases t.

The following lines contain t test cases. One test case per line.

Each test case is in the format A S B Q C, where A,B and C are the numbers, S is the sign between A and B and Q is the '=' sign.